// Qanat.Web/src/app/shared/generated/model/water-account-request-changes-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { WaterAccountRequestChangesParcelDto } from '././water-account-request-changes-parcel-dto';

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class WaterAccountRequestChangesDto { 
    WaterAccountID?: number;
    WaterAccountName?: string;
    WaterAccountNumber?: number;
    ContactName?: string;
    ContactAddress?: string;
    Parcels?: Array<WaterAccountRequestChangesParcelDto>;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface WaterAccountRequestChangesDtoForm { 
    WaterAccountID?: FormControl<number>;
    WaterAccountName?: FormControl<string>;
    WaterAccountNumber?: FormControl<number>;
    ContactName?: FormControl<string>;
    ContactAddress?: FormControl<string>;
    Parcels?: FormControl<Array<WaterAccountRequestChangesParcelDto>>;
}

export class WaterAccountRequestChangesDtoFormControls { 
    public static WaterAccountID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WaterAccountName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WaterAccountNumber = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ContactName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ContactAddress = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Parcels = (value: FormControlState<Array<WaterAccountRequestChangesParcelDto>> | Array<WaterAccountRequestChangesParcelDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<WaterAccountRequestChangesParcelDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}