// Qanat.Web/src/app/shared/generated/model/water-account-request-changes-parcel-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ZoneDisplayDto } from '././zone-display-dto';

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class WaterAccountRequestChangesParcelDto { 
    ParcelID?: number;
    ParcelNumber?: string;
    WaterAccountID?: number;
    WaterAccountName?: string;
    AllocationZone?: ZoneDisplayDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface WaterAccountRequestChangesParcelDtoForm { 
    ParcelID?: FormControl<number>;
    ParcelNumber?: FormControl<string>;
    WaterAccountID?: FormControl<number>;
    WaterAccountName?: FormControl<string>;
    AllocationZone?: FormControl<ZoneDisplayDto>;
}

export class WaterAccountRequestChangesParcelDtoFormControls { 
    public static ParcelID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ParcelNumber = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WaterAccountID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WaterAccountName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AllocationZone = (value: FormControlState<ZoneDisplayDto> | ZoneDisplayDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<ZoneDisplayDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}